Convert ambiguous date into ISO standard format.
Sorry to be picky, but I got bitten by some ambiguous date format problems recently so...

The date I am converting (9/6/19) is Sep 6th if it is in US format and is June 9th for all other English speaking countries (I'm highlighting English countries only because this text is in English).  I determined that it is US format by looking at the date of the commit.

I suggest that stuff like this use [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format because it is unambiguous, it is popular amongst developers, it is getting wider adoption, and it is a great format!  Thanks.