Fix major version check
There is existing functionality which is supposed to allow you to specify a major version e.g `nvm install 9` and have it install the latest v9. This was just checking the length of version is 1, which means `nvm install 10` installs `10.0.0`. I've changed this to check if the string contains no `.`, which (I think?) is a more reasonable approach which will work for future versions.